Revert "Refactor app options"
This reverts commit 8991f33f0e9c31c1af140383e8e93e8ef9a48146.
