+++ cli/azd/cmd/container.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"io"
+	"net/http"
 	"os"
 
 	"github.com/azure/azure-dev/cli/azd/cmd/actions"
@@ -86,26 +87,32 @@ func resolveAction[T actions.Action](container *ioc.NestedContainer, actionName 
 }
 
 // Registers common Azd dependencies
-func registerCommonDependencies(options AppOptions, container *ioc.NestedContainer) {
+func registerCommonDependencies(container *ioc.NestedContainer) {
 	container.RegisterSingleton(output.GetCommandFormatter)
 
 	container.RegisterSingleton(func(
 		rootOptions *internal.GlobalCommandOptions,
-		formatter output.Formatter) input.Console {
-		writer := options.ConsoleHandles.Stdout
+		formatter output.Formatter,
+		cmd *cobra.Command) input.Console {
+		writer := cmd.OutOrStdout()
 		// When using JSON formatting, we want to ensure we always write messages from the console to stderr.
 		if formatter != nil && formatter.Kind() == output.JsonFormat {
-			writer = options.ConsoleHandles.Stderr
+			writer = cmd.ErrOrStderr()
 		}
 
 		if os.Getenv("NO_COLOR") != "" {
 			writer = colorable.NewNonColorable(writer)
 		}
 
-		isTerminal := options.ConsoleHandles.Stdin == os.Stdin && isatty.IsTerminal(os.Stdin.Fd()) &&
-			options.ConsoleHandles.Stdout == os.Stdout && isatty.IsTerminal(os.Stdout.Fd())
+		isTerminal := cmd.OutOrStdout() == os.Stdout &&
+			cmd.InOrStdin() == os.Stdin && isatty.IsTerminal(os.Stdin.Fd()) &&
+			isatty.IsTerminal(os.Stdout.Fd())
 
-		return input.NewConsole(rootOptions.NoPrompt, isTerminal, writer, options.ConsoleHandles, formatter)
+		return input.NewConsole(rootOptions.NoPrompt, isTerminal, writer, input.ConsoleHandles{
+			Stdin:  cmd.InOrStdin(),
+			Stdout: cmd.OutOrStdout(),
+			Stderr: cmd.ErrOrStderr(),
+		}, formatter)
 	})
 
 	container.RegisterSingleton(func(console input.Console, rootOptions *internal.GlobalCommandOptions) exec.CommandRunner {
@@ -119,7 +126,7 @@ func registerCommonDependencies(options AppOptions, container *ioc.NestedContain
 	})
 	container.RegisterSingleton(input.NewConsoleMessaging)
 
-	container.RegisterSingleton(func() httputil.HttpClient { return options.HttpClient })
+	container.RegisterSingleton(func() httputil.HttpClient { return &http.Client{} })
 
 	// Auth
 	container.RegisterSingleton(auth.NewLoggedInGuard)
@@ -322,7 +329,7 @@ func registerCommonDependencies(options AppOptions, container *ioc.NestedContain
 	}
 
 	// Other
-	container.RegisterSingleton(func() clock.Clock { return options.Clock })
+	container.RegisterSingleton(clock.New)
 
 	// Service Targets
 	serviceTargetMap := map[project.ServiceTargetKind]any{
