+++ cli/azd/test/functional/initialize_test.go
@@ -55,7 +55,7 @@ func Test_CommandsAndActions_Initialize(t *testing.T) {
 
 	// Creates the azd root command with a "Skip" middleware that will skip the invocation
 	// of the underlying command / actions
-	rootCmd := cmd.NewRootCmd(true, chain, nil)
+	rootCmd := cmd.NewRootCmd(true, chain)
 	testCommand(t, rootCmd, ctx, chain, tempDir)
 }
 
@@ -77,7 +77,7 @@ func testCommand(
 			fullCmd := fmt.Sprintf("%s %s", testCmd.Parent().CommandPath(), use)
 			args := strings.Split(fullCmd, " ")[1:]
 			args = append(args, "--cwd", cwd)
-			childCmd := cmd.NewRootCmd(true, chain, nil)
+			childCmd := cmd.NewRootCmd(true, chain)
 			childCmd.SetArgs(args)
 			err := childCmd.ExecuteContext(ctx)
 			require.NoError(t, err)
